Crate platform_data

Macros

query

Structs

AddrToRaw

Hybrid

LinksConstants

Point

A structure representing a single value repeated multiple times.

PointIter

Iterator for Point that yields copies of the stored value.

Query

RawToAddr

Enums

Error

Flow

Represents the control flow of an operation, similar to ControlFlow.

Traits

LinkReference

A composite trait for types that can be used as link identifiers.

Links

ToQuery

Type Aliases

ReadHandler

WriteHandler